Increased Systematic Color Palette
I've decided to increase the Systematic color palette from 64 to 70.  I've modified the existing yellow colors and added turquoise and warm gray swatches.  That allowed me to create a PlayStation and GameCube dev kit, and bring some warmth to some classic systems that are beige, but have previously been released in gray.